websocket/server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[silero,websocket,google]>=0.0.82
loguru
